requirements.txt
git+https://github.com/Rapptz/discord.py
PyNaCl==1.3.0
dnspython==1.16.0
aiohttp>=3.6.0,<3.7.0
websockets>=6.0,!=7.0,!=8.0,!=8.0.1,<9.0
discord.py[voice]==1.2.5
pip
youtube_dl
requests
discord.py
